# Conflicts:
#	.gitlab-ci.yml
#	xplat-meta-oqsengine-boot/src/main/java/com/xforceplus/ultraman/oqsengine/boot/config/CommonConfiguration.java
#	xplat-meta-oqsengine-boot/src/main/java/com/xforceplus/ultraman/oqsengine/boot/config/MetaManagerConfiguration.java
#	xplat-meta-oqsengine-boot/src/main/java/com/xforceplus/ultraman/oqsengine/boot/grpc/service/EntityRebuildServiceOqs.java
#	xplat-meta-oqsengine-boot/src/main/java/com/xforceplus/ultraman/oqsengine/boot/rest/DevOpsController.java
#	xplat-meta-oqsengine-boot/src/main/resources/application.yaml
#	xplat-meta-oqsengine-boot/src/test/java/com/xforceplus/ultraman/oqsengine/core/service/integration/JoinSelectTest.java
#	xplat-meta-oqsengine-boot/src/test/java/com/xforceplus/ultraman/oqsengine/core/service/integration/UserCaseTest.java
#	xplat-meta-oqsengine-cdc/src/test/java/com/xforceplus/ultraman/oqsengine/cdc/benchmark/BigBatchSyncTest.java
#	xplat-meta-oqsengine-cdc/src/test/java/com/xforceplus/ultraman/oqsengine/cdc/benchmark/MassageUnpackBenchmarkTest.java
#	xplat-meta-oqsengine-cdc/src/test/java/com/xforceplus/ultraman/oqsengine/cdc/cdcerrors/CdcErrorStorageTest.java
#	xplat-meta-oqsengine-cdc/src/test/java/com/xforceplus/ultraman/oqsengine/cdc/consumer/ConsumerErrorTest.java
#	xplat-meta-oqsengine-cdc/src/test/java/com/xforceplus/ultraman/oqsengine/cdc/consumer/impl/SphinxConsumerServiceTest.java
#	xplat-meta-oqsengine-changelog/src/test/java/com/xforceplus/ultraman/oqsengine/changelog/ReplayServiceTest.java
#	xplat-meta-oqsengine-changelog/src/test/java/com/xforceplus/ultraman/oqsengine/changelog/config/ChangelogConfiguration.java
#	xplat-meta-oqsengine-common/src/main/java/com/xforceplus/ultraman/oqsengine/common/watch/RedisLuaScriptWatchDog.java
#	xplat-meta-oqsengine-index/src/main/java/com/xforceplus/ultraman/oqsengine/storage/index/sphinxql/SphinxQLManticoreIndexStorage.java
#	xplat-meta-oqsengine-index/src/main/java/com/xforceplus/ultraman/oqsengine/storage/index/sphinxql/helper/SphinxQLHelper.java
#	xplat-meta-oqsengine-index/src/test/java/com/xforceplus/ultraman/oqsengine/storage/index/sphinxql/SphinxQLManticoreIndexStorageSearchTest.java
#	xplat-meta-oqsengine-index/src/test/java/com/xforceplus/ultraman/oqsengine/storage/index/sphinxql/SphinxQLManticoreIndexStorageTest.java
#	xplat-meta-oqsengine-index/src/test/java/com/xforceplus/ultraman/oqsengine/storage/index/sphinxql/helper/SphinxQLHelperTest.java
#	xplat-meta-oqsengine-index/src/test/java/com/xforceplus/ultraman/oqsengine/storage/index/sphinxql/strategy/condition/select/MatchConditionBuilderTest.java
#	xplat-meta-oqsengine-metadata/src/main/java/com/xforceplus/ultraman/oqsengine/metadata/StorageMetaManager.java
#	xplat-meta-oqsengine-metadata/src/main/java/com/xforceplus/ultraman/oqsengine/metadata/cache/DefaultCacheExecutor.java
#	xplat-meta-oqsengine-metadata/src/main/java/com/xforceplus/ultraman/oqsengine/metadata/executor/EntityClassSyncExecutor.java
#	xplat-meta-oqsengine-metadata/src/test/java/com/xforceplus/ultraman/oqsengine/metadata/MockerRequestClientHelper.java
#	xplat-meta-oqsengine-metadata/src/test/java/com/xforceplus/ultraman/oqsengine/metadata/executor/EntityClassSyncExecutorTest.java
#	xplat-meta-oqsengine-metadata/src/test/java/com/xforceplus/ultraman/oqsengine/metadata/utils/FileReaderUtilsTest.java
#	xplat-meta-oqsengine-metadata/src/test/resources/5_2_0.json
#	xplat-meta-oqsengine-pojo/src/main/java/com/xforceplus/ultraman/oqsengine/pojo/dto/values/DecimalValue.java
#	xplat-meta-oqsengine-pojo/src/test/java/com/xforceplus/ultraman/oqsengine/pojo/dto/values/DecimalValueTest.java
#	xplat-meta-oqsengine-status/src/main/java/com/xforceplus/ultraman/oqsengine/status/impl/CommitIdStatusServiceImpl.java
#	xplat-meta-oqsengine-testcontainer/src/main/java/com/xforceplus/ultraman/oqsengine/testcontainer/container/ContainerStarter.java
#	xplat-meta-oqsengine-testcontainer/src/main/java/com/xforceplus/ultraman/oqsengine/testcontainer/junit4/ContainerRunner.java
#	xplat-meta-oqsengine-tokenizer/src/main/java/com/xforceplus/ultraman/oqsengine/tokenizer/Tokenizer.java
#	xplat-meta-oqsengine-tokenizer/src/main/java/com/xforceplus/ultraman/oqsengine/tokenizer/segmentation/JcsegTokenizer.java
#	xplat-meta-oqsengine-tokenizer/src/test/java/com/xforceplus/ultraman/oqsengine/tokenizer/DefaultTokenizerFactoryTest.java
#	xplat-meta-oqsengine-tokenizer/src/test/java/com/xforceplus/ultraman/oqsengine/tokenizer/WrapperTokenizerTest.java
#	xplat-meta-oqsengine-tokenizer/src/test/resources/jcseg.properties
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-client/pom.xml
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-client/src/main/java/com/xforceplus/ultraman/oqsengine/meta/executor/RequestWatchExecutor.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-client/src/main/java/com/xforceplus/ultraman/oqsengine/meta/handler/DoNothingRequestHandler.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-client/src/test/java/com/xforceplus/ultraman/oqsengine/meta/BaseTest.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-common/src/main/java/com/xforceplus/ultraman/oqsengine/meta/common/constant/Constant.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-common/src/main/java/com/xforceplus/ultraman/oqsengine/meta/common/dto/AbstractWatcher.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-common/src/main/java/com/xforceplus/ultraman/oqsengine/meta/common/dto/IWatcher.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-common/src/main/java/com/xforceplus/ultraman/oqsengine/meta/common/dto/WatchElement.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-common/src/main/java/com/xforceplus/ultraman/oqsengine/meta/common/exception/Code.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-common/src/main/java/com/xforceplus/ultraman/oqsengine/meta/common/metrics/ConnectorMetricsDefine.java
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-server/pom.xml
#	xplat-meta-oqsengine-transfer/xplat-meta-oqsengine-meta-springboot-starter/pom.xml

+++ xplat-meta-oqsengine-devops/src/main/java/com/xforceplus/ultraman/oqsengine/devops/rebuild/model/DevOpsTaskInfo.java
@@ -43,7 +43,7 @@ public interface DevOpsTaskInfo {
      *
      * @return 百分比0-100.
      */
-    int getProgressPercentage();
+    long getProgressPercentage();
 
     long getMaintainid();
 
@@ -57,7 +57,7 @@ public interface DevOpsTaskInfo {
 
     IEntityClass getEntityClass();
 
-    int getBatchSize();
+    long getBatchSize();
 
     int getFinishSize();
 
@@ -73,7 +73,7 @@ public interface DevOpsTaskInfo {
 
     long startId();
 
-    void setBatchSize(int size);
+    void setBatchSize(long size);
 
     int failedRecovers();
